LayoutTests/imported/w3c/web-platform-tests/speculation-rules/speculation-tags/prefetch-eagerness-pointer-hover.https-expected.txt


FAIL Sec-Speculation-Tags prefetch with eagerness triggered by point hover assert_equals: Sec-Speculation-Tags expected "\"moderate\"" but got "conservative"
